requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.10.5
certvalidator==0.11.1
cffi==1.17.1
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
crypto==1.4.1
cryptography==45.0.5
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
deepface==0.0.93
defusedxml==0.7.1
distro==1.9.0
easyocr==1.7.2
endesive==2.19.0
executing==2.2.1
filelock==3.19.1
fire==0.7.0
flasgger==0.9.7.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.9.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.73.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
imutils==0.5.4
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
keybert==0.9.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.8
lazy_loader==0.4
libclang==18.1.1
lxml==6.0.0
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.6.3
mypy_extensions==1.1.0
Naked==0.1.32
namex==0.1.0
networkx==3.5
ninja==1.13.0
nltk==3.9.1
numpy==2.3.3
openai==1.97.1
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
orjson==3.11.0
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
paramiko==3.5.1
parso==0.8.5
PassportEye==2.2.2
pdfminer==20191125
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
playwright==1.54.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycairo==1.28.0
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.1
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyKCS11==1.5.17
PyMuPDF==1.26.3
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==5.9.0
pypdfium2==4.30.0
PySocks==1.7.1
pytesseract==0.3.13
python-barcode==0.15.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pkcs11==0.8.1
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
retina-face==0.0.17
rich==14.0.0
rpds-py==0.26.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
selenium==4.34.2
sentence-transformers==5.1.1
shapely==2.1.1
shellescape==3.8.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
supervision==0.25.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tifffile==2025.9.9
tokenizers==0.22.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uharfbuzz==0.50.2
ultralytics==8.3.155
ultralytics-thop==2.0.14
uritools==5.0.0
urllib3==2.5.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wfastcgi==3.0.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zstandard==0.23.0
